+++ docs/reference/changelog.rst
@@ -9,6 +9,7 @@ Develop
 -----------------
 * [ENHANCEMENT] Add ``.feather`` file support to PandasDatasource
 * [FEATURE] A profiler that builds suites from JSONSchema files.
+* [FEATURE] Add new expectation expect_table_row_count_to_equal_other_table for SqlAlchemyDataset
 
 0.11.5
 -----------------